Big refactoring at the way the charts were requesting and receiving data -> the logic is being isolated in a way that each chart type is a component itself /=> we can load different views depending on the type of chart selected/. A chart component is also responsible for building up the query after selecting its respective properties and passing it on to a chart service, which will get the latest records from ElasticSearch and with the data returned the service will call the specific chart module to build the EChartOption. The dashboard-element.component is listening for any changes coming from the service and updates itself accordingly.
Fixed a bug that kept the query property of every newly added element listening to the service changes, making all the newly added elements with the same query. Added the possibility of including filters in a bar chart query.

When editing a previously added chart element, all the properties are preselected. After selecting a different ones they will also update.

In the menu component we subscribe to the router, so the moment a url changes we will default the dashboard name to null as well as a boolean /= false/ that controls the “Add dashboard” element to appear instead of an input.